+++ internal/e2e-js/tests/v2Webrtc/webrtcCalling.spec.ts
@@ -1,8 +1,15 @@
-import { test, expect, Page } from '../fixtures'
+import {
+  expect,
+  Page,
+  test
+} from '../../fixtures'
+
 import {
   SERVER_URL,
   createTestJWTToken,
-} from '../utils'
+  expectRelayConnected,
+  expectv2TotalAudioEnergyToBeGreaterThan
+} from '../../utils'
 	      
 test.describe('V2Calling', () => {
   test('should handle one webrtc endpoint calling to a second webrtc endpoint waiting to answer', async ({
@@ -37,33 +44,9 @@ test.describe('V2Calling', () => {
     const jwtCallee = await createTestJWTToken({ 'resource': 'vanilla-callee' })
     expect(jwtCallee).not.toBe(null)
 
-    const expectRelayConnected = async (page: Page, jwt: string) => {
-      // Project locator
-      const project = page.locator('#project')
-      expect(project).not.toBe(null)
-
-      // Token locator
-      const token = page.locator('#token')
-      expect(token).not.toBe(null)
-
-      // Populate project and token using locators
-      await project.fill(envRelayProject)
-      await token.fill(jwt)
-
-      // Click the connect button, which calls the connect function in the browser
-      await page.click('#btnConnect')
-
-      // Start call button locator
-      const startCall = page.locator('#startCall')
-      expect(startCall).not.toBe(null)
-
-      // Wait for call button to be enabled when signalwire.ready occurs
-      await expect(startCall).toBeEnabled()
-    }
-
     // Wait for both caller and callee to get connected to Relay
-    await expectRelayConnected(pageCaller, jwtCaller)
-    await expectRelayConnected(pageCallee, jwtCallee)
+    await expectRelayConnected(pageCaller, envRelayProject, jwtCaller)
+    await expectRelayConnected(pageCallee, envRelayProject, jwtCallee)
 
     const expectCallStarted = async (page: Page, to: string, from: string) => {
       // To Number locator
@@ -122,68 +105,12 @@ test.describe('V2Calling', () => {
     await expectVideoMediaStreams(pageCaller)
     await expectVideoMediaStreams(pageCallee)
 
-    const expectTotalAudioEnergyToBeGreaterThan = async (
-      page: Page,
-      value: number
-    ) => {
-      const audioStats = await page.evaluate(async () => {
-        // @ts-expect-error
-        const currentCall = window.__currentCall
-        // @ts-expect-error
-        const audioReceiver = currentCall.peer.instance.getReceivers().find(r => r.track.kind === 'audio')
-
-        const audioTrackId = audioReceiver.track.id
-    
-        const stats = await currentCall.peer.instance.getStats(null)
-        const filter = {
-          'inbound-rtp': [
-            'audioLevel',
-            'totalAudioEnergy',
-            'totalSamplesDuration',
-            'totalSamplesReceived',
-            'packetsDiscarded',
-            'lastPacketReceivedTimestamp',
-            'bytesReceived',
-            'packetsReceived',
-            'packetsLost',
-            'packetsRetransmitted',
-          ],
-        }
-        const result: any = {}
-        Object.keys(filter).forEach((entry) => {
-          result[entry] = {}
-        })
-    
-        stats.forEach((report: any) => {
-          for (const [key, value] of Object.entries(filter)) {
-            if (
-              report.type == key &&
-              report['mediaType'] === 'audio' &&
-              report['trackIdentifier'] === audioTrackId
-            ) {
-              value.forEach((entry) => {
-                if (report[entry]) {
-                  result[key][entry] = report[entry]
-                }
-              })
-            }
-          }
-        }, {})
-    
-        return result
-      })
-      console.log('audioStats', audioStats)
-    
-      expect(audioStats['inbound-rtp']['totalAudioEnergy']).toBeGreaterThan(value)
-    }
-
     // Give some time to collect audio from both pages
     await pageCaller.waitForTimeout(10000)
 
     // Check the audio energy level is above threshold
-    await expectTotalAudioEnergyToBeGreaterThan(pageCaller, 0.4)
-    await expectTotalAudioEnergyToBeGreaterThan(pageCallee, 0.4)
-
+    await expectv2TotalAudioEnergyToBeGreaterThan(pageCaller, 0.4)
+    await expectv2TotalAudioEnergyToBeGreaterThan(pageCallee, 0.4)
 
     // Click the caller hangup button, which calls the hangup function in the browser
     await pageCaller.click('#hangupCall')
